+++ Changelog.md
@@ -6,6 +6,24 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 
+## [3.0.13] - 2025-01-06
+
+### Fixed
+
+- [Active queues] Decreased the amount of data that gets refreshed, resulting in a small performance bump.
+
+- [Active queues] The cancel button will not disappear anymore when you have had a manually listed group in the group finder.
+
+- [DropChecker] The search for items will now start at a single character entered instead of two.
+
+- [Options] The "Classic" background option is once again available.
+
+- [ApplicationViewer] The app viewer won't jump into the foreground anymore if the leader edited the listing / relisted the group.
+
+Renamed the "Statistics" tab in the mainframe to "Progress" since there's a lot of change incoming over the next two months and this fits more :).
+
+
+
 ## [3.0.12] - 2024-12-27
 
 ### Changed
@@ -2974,7 +2992,8 @@ Basically just the rating for the bracket and the tier they're currently in
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[unreleased]: https://github.com/NintendoLink07/MythicIOGrabber/compare/3.0.12..HEAD
+[unreleased]: https://github.com/NintendoLink07/MythicIOGrabber/compare/3.0.13..HEAD
+[3.0.13]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/3.0.13
 [3.0.12]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/3.0.12
 [3.0.11]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/3.0.11
 [3.0.10]: https://github.com/NintendoLink07/MythicIOGrabber/releases/tag/3.0.10